Add controllers and services

+++ Ms.Inventory.BusinessLogic.Blo/Inventory/InventoryDataBlo.cs
@@ -6,7 +6,7 @@ namespace Ms.Inventory.BusinessLogic.Blo.Inventory
     public class InventoryDataBlo
     {
         public string InventoryId { get; set; }
-        public long ItemReference { get; set; }
+        public int ItemReference { get; set; }
         public string Location { get; set; }
         public DateTimeOffset DateOfInventory { get; set; }
         public HashSet<string> Tags { get; set; }
