+++ Ms.Inventory.Dto/Inventory/InventoryDataDto.cs
@@ -8,7 +8,7 @@ namespace Ms.Inventory.Dto.Inventory
     {
         [StringLength(32, ErrorMessage = "Inventory Id should not be longer than 32 characters")]
         public string InventoryId { get; set; }
-        public long ItemReference { get; set; } 
+        public int ItemReference { get; set; } 
         public string Location { get; set; }
         public DateTimeOffset DateOfInventory { get; set; }
         public HashSet<string> Tags { get; set; }
